lala

+++ app/views.py
@@ -22,13 +22,15 @@ def borrarauto():
 	if request.method == 'POST':
 		patente =  request.form['PATENTE']
 		try:
-			sql= """ delete from mediciones where mediciones.patente= %s"""%(patente)
-			cur.execute(sql)
-			sql= """ delete from autos where mediciones.patente = %s""" %(patente)
-			cur.execute(sql)
-			conn.commit()
+			sql= """delete from mediciones where mediciones.patente='%s';"""%(patente)
+			print(sql)
+			print(cur.execute(sql))
+			sql= """delete from autos where autos.patente ='%s';""" %(patente)
+			print(cur.execute(sql))
+			print(sql)
 			return render_template("eliminacion_ok.html",nombre="nombre")
 		except:
+			print("error")
 			return render_template("eliminar_error.html",nombre="nombre")
 	else:
 		return render_template("eliminar_auto.html",nombre="nombre")
@@ -39,11 +41,17 @@ def borrarauto():
 def borrarcliente():
 	if request.method == 'POST':
 		rutcliente = request.form['RUT']
+		print("se procede a eliminar el cliente",rutcliente)
 		try:
+			print("sql lo intenta")
 			sql= """ delete from clientes where clientes.rut = %s"""%(rutcliente)
-			cur.execute(sql)
+			x=cur.execute(sql)
+			print(x)
+			if x is None:
+				print("sql falla")
+				return render_template("eliminar_error.html",nombre="nombre")
 			sql= """ delete from autos where autos.rut = %s"""%(rutcliente)
-			cur.execute(sql)
+			print(cur.execute(sql))
 			conn.commit()
 			return render_template("eliminacion_ok.html",nombre="nombre")
 		except:
@@ -71,7 +79,7 @@ def borrardebe():
 def actualizar():
 	return render_template('actualizar_inicio.html')
 
-@app.route(/'ACTUALIZAR_DUENO',methods=['GET','POST'])
+@app.route('/ACTUALIZAR_DUENO',methods=['GET','POST'])
 def actualizardueno():
 	if request.method == 'POST':
 		rutdueno = resquest.form['RUT']
@@ -87,7 +95,7 @@ def actualizardueno():
 		return render_template("actualizar_dueno.html",nombre="nombre")		
 
 
-@app.route(/'ACTUALIZAR_FECHADEBE',methods=['GET','POST'])
+@app.route('/ACTUALIZAR_FECHADEBE',methods=['GET','POST'])
 def actualizardebe():
 	if request.method == 'POST':
 		rutdebe = resquest.form['RUT']
@@ -102,7 +110,7 @@ def actualizardebe():
 	else:
 		return render_template("actualizar_dueno.html",nombre="nombre")	
 
-@app.route(/'ACTUALIZAR_TELEFONO',methods=['GET','POST'])
+@app.route('/ACTUALIZAR_TELEFONO',methods=['GET','POST'])
 def actualizartelefono():
 	if request.method == 'POST':
 		ruttel = resquest.form['RUT']
@@ -117,11 +125,11 @@ def actualizartelefono():
 	else:
 		return render_template("actualizar_dueno.html",nombre="nombre")	
 
-@app.route(/'CREAR',methods=['GET','POST'])
+@app.route('/CREAR',methods=['GET','POST'])
 def crear():
 	return render_template('crear_inicio.html')
 
-@app.route(/'CREAR_SENSOR',methods=['GET','POST'])
+@app.route('/CREAR_SENSOR',methods=['GET','POST'])
 def crearsensor():
 	if request.method == 'POST':
 		idsensor = resquest.form['ID_SENSOR']
